requirements.txt
numpy
pandas
scikit-learn
ipykernel
jupyter
matplotlib
tensorflow
tqdm
python-dotenv